Working basic demo: all collisions
All collisions work, this includes walls, sticks, and even
game restart and add point to winner when ball misses the
sticks. Sticks can move and reflect the ball.

TODO: Impliment score on the display, and change so that
when the ball hits a stick, its y velocity will be affected
by velocity of the stick (currently just reflects). I think this
will be done using two rags/flags and XOR to see if the "has touched
stick" flag has changed.